# adding the feeling good and history logic

+++ database/migrations/2023_01_04_070632_create_luck_histories_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('luck_histories', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('score');
+            $table->decimal('score', 10, 2);
             $table->timestamps();
         });
     }
